delete netlify config
Have decided to move this blog into Cloudflare.

Reason for doing so that is that cloudflare publishes updates to
the `master` branch of this repository on github. i.e. I can
glance at the repo, see the build status and immediately know
that not only is the site being built, but that it's being done
on Cloudlfare.

Previously, this blog was hosted on DigitalOcean ... and I forgot
about it. Took a day or two before finally tracking down where
the pages were being served from. And then some more time to
fix the failing build pipelines. The lack of build status updates
on github was part of the problem - IMO.

Netlify does do biuld status - but only on PRs it seems. Not to
the master branch. And I need the status easily visible.